# openapi/sovereign.v1.yaml
openapi: 3.1.0
info:
  title: Sovereign REST API
  version: "1.0.0"
  description: |
    Control-plane API for Sovereign tenants. Hand-written source of truth
    (D-5): the route set in internal/api/router.go must stay in parity with
    this spec in both directions, enforced by the OpenAPI drift test.
servers:
  - url: https://id.<domain>/api/v1
    description: Identity host, versioned API base.
# Default security for the API: bearer (programmatic) or session cookie
# (browser). Operations that are anonymous override this with `security: []`.
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      summary: List all users across tenants (instance admin, scope admin:users:read)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paged users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    maxItems: 1000
                    items:
                      $ref: "#/components/schemas/User"
                  offset:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      operationId: adminCreateUser
      summary: Create a user and send an invite (instance admin, scope admin:users:write, idempotent)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created and invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        detail:
          type: string
    User:
      type: object
      required: [id, handle]
      properties:
        id:
          type: string
        handle:
          type: string
        display_name:
          type: string
        is_admin:
          type: boolean
...
    CreateUserRequest:
      type: object
      required: [handle, email]
      properties:
        tenant_id:
          type: string
        handle:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        is_admin:
          type: boolean
...